Fix blogpost sorting (#311)
* Switched to unfiltered commit

2016 posts were being removed previously; just switching the data back
to the unfiltered post array.

* Fixed bug by modifying sort

Previously, the post dates were being compared as so:
    postA.date < postB.date
This was causing the sorting problems we saw (for reasons unknown to
me). I changed it to
    new Date(postB.date) - new Date(postA.date)
so as to a) use JavaScript's built in Date object and b) use a more
direct way of determining which Date is higher (as the implementation of
a less-than or greater-than operator is more uncertain than that of a
subtraction operator).